refactored processing groups in utils.js
