+++ src/js/utils.js
@@ -316,3 +316,20 @@ export function filterAndSumGeoCounts(childData, primaryGroup, selectedSubindica
 
     return sumData;
 }
+
+export function getFilterGroups(groups, primaryGroup){
+    groups = groups.reduce(function (memo, e1) {
+        let matches = memo.filter(function (e2) {
+            return e1.name === e2.name
+        })
+        if (matches.length == 0)
+            memo.push(e1)
+        return memo;
+    }, [])
+
+    groups = groups.filter((g) => {
+        return g.name !== primaryGroup
+    });
+
+    return groups;
+}
